Prevent infinite SPEAKER_OFF looping
Originating from an external request to turn speaker off, if the routing
ends up being the same in the case of a video call, then prevent the
additional SPEAKER_OFF msg from being sent. This normally ends up being
needed for the pending code path to turn off the notification for
instance but if the routing doesn't change, then we should not send this
extra message.

Bug: 425988954
Flag: EXEMPT bug fix
Test: atest CallAudioRouteControllerTest

Signed-off-by: Pranav Temkar <pranavtemkar@gmail.com>
Change-Id: I3ded7bf357ff2b8362e9ce4997cadcb6261b5073